(fix) Fix invalid date error in laboratory dashboard (#124)
Fixes an issue where the date range picker in the laboratory dashboard was throwing an error because the default date format for the Carbon date picker is `m/d/Y` while the date range picker was providing dates in the format `d/m/Y`. Explicitly setting the date format to `d/m/Y` fixes the issue.

Other unrelated changes in the diff include:

- Removing the `omrs-main-content` class from the laboratory dashboard, as it doesn't do anything useful
- Reordering imports for easier reading
- Replacing the date range picker's `onClose` prop with an `onChange` prop, allowing date range changes to be reflected in the UI immediately